[DEV-1374, 1377, 1380] Consent WS API rev 1
The first version of the Consent API, including web services for creating, retrieving, and updating consents,
as well as an autosuggest service for ontology terms (and an ontology indexing service that uses Elastic Search).

This implementation borrows some code (including the models package) from the datause-prototype repository,
and implements the rest using a Dropwizard template adapted from the work on the BOSS API.

ConsentResource is the main per-consent Web Service endpoint.  ConsentAPI is an interface modeling the primary
actions in the service, and ConsentDAO is the implementation of the database backend.  This implementation currently
stores all its backing information in a database configured using the Dropwizard YAML.